# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /m365/groups:
    get:
      operationId: getM365Groups
      tags: [m365]
      summary: Entra ID groups list
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Groups response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/M365GroupsData"
  /m365/groups/with-metadata:
    get:
      operationId: getM365GroupsWithMetadata
      tags: [m365]
      summary: Entra ID groups list with field metadata
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Groups response with field metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/M365GroupsDataWithMetadata"
...
  /m365/groups/{id}/device-members:
    get:
      operationId: getM365GroupDeviceMembers
      tags: [m365]
      summary: Display names of every device object in a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device display names in the group
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceNames:
                    type: array
                    items:
                      type: string
                required: [deviceNames]
  /m365/export/findings.csv:
    get:
      operationId: getM365ExportFindingsCsv
      tags: [m365]
      summary: Findings register as CSV evidence
      parameters:
        - $ref: "#/components/parameters/ExportScanId"
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /m365/export/findings.xlsx:
    get:
      operationId: getM365ExportFindingsXlsx
      tags: [m365]
      summary: Findings register as an Excel workbook
      parameters:
        - $ref: "#/components/parameters/ExportScanId"
      responses:
        "200":
          description: Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /m365/export/evidence.xlsx:
    get:
      operationId: getM365ExportEvidenceXlsx
      tags: [m365]
      summary: Evidence pack as an Excel workbook
      parameters:
        - $ref: "#/components/parameters/ExportScanId"
      responses:
        "200":
          description: Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /m365/export/executive.html:
    get:
      operationId: getM365ExportExecutiveHtml
      tags: [m365]
      summary: Executive posture report as print-ready HTML
      parameters:
        - $ref: "#/components/parameters/ExportScanId"
      responses:
        "200":
          description: Self-contained HTML document
          content:
            text/html:
              schema:
                type: string
  /m365/export/executive.pdf:
    get:
      operationId: getM365ExportExecutivePdf
      tags: [m365]
      summary: Executive posture report as PDF
      parameters:
        - $ref: "#/components/parameters/ExportScanId"
      responses:
        "200":
          description: PDF document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
components:
  parameters:
    ExportScanId:
      name: scanId
      in: query
      required: false
      description: Archived scan to export; omit for the live register.
      schema:
        type: string
  schemas:
...
    EvidenceStatus:
      type: string
      enum: [apiBacked, partial, manual, automationCandidate, notAssessed]
    ConfidenceLabel:
      type: string
      enum: [high, medium, low, unknown]
    FieldEvidenceMetadata:
      type: object
      properties:
        evidenceStatus:
          $ref: "#/components/schemas/EvidenceStatus"
        confidenceLabel:
          $ref: "#/components/schemas/ConfidenceLabel"
        sourceLabel:
          type: string
        notes:
          type: array
          items:
            type: string
      required: [evidenceStatus, confidenceLabel]
    FieldMetadataMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/FieldEvidenceMetadata"
...
    GroupItem:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        description:
          type: string
          nullable: true
        groupTypes:
          type: array
          items:
            type: string
        securityEnabled:
          type: boolean
      required: [id, displayName, groupTypes, securityEnabled]
    M365GroupsData:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupItem"
      required: [groups]
    M365GroupsDataWithMetadata:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/M365GroupsData"
        fieldMetadata:
          $ref: "#/components/schemas/FieldMetadataMap"
        metadataVersion:
          type: string
      required: [data, fieldMetadata, metadataVersion]
...